backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=42.0.0",
    "didkit>=0.3.3",
    "fastapi>=0.111.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.3.0",
    "pydantic[email]>=2.7.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "supabase>=2.4.0",
    "uvicorn[standard]>=0.29.0",
]
